Project_Starter_Files-Building_a_Metrics_Dashboard/reference-app/backend/requirements.txt
Flask==1.1.2
requests==2.23.0
pylint==2.4.0
gunicorn==20.0.4
pymongo==3.11.0
Flask-PyMongo==2.3.0
prometheus-client==0.12.0
Werkzeug==0.15
Jinja2==2.10.1
itsdangerous==0.24
click==5.1
markupsafe==2.0.1
prometheus-flask-exporter==0.18.1

opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-exporter-jaeger==1.17.0
opentelemetry-instrumentation-flask==0.38b0
